refactor: Extracted creation of products business logic
