+++ app/src/apps/products/hooks/index.ts
@@ -1,0 +1,1 @@
+export { useCreateProduct } from './use-create-product'
