+++ app/src/apps/products/views/CreateProduct.vue
@@ -14,47 +14,13 @@
 </template>
 
 <script lang="ts" setup>
-import { ProductInput, ProductStatus } from '@/API'
+import { useCreateProduct } from '@/apps/products/hooks'
 import {
   InputGroup,
   SelectGroup,
   TextAreaGroup,
 } from '@/components/core-ui/fields'
 import { Button } from '@/components/core-ui/buttons'
-import { createProduct } from '@/graphql/mutations'
-import { API, graphqlOperation } from 'aws-amplify'
-import { reactive, toRefs } from 'vue'
 
-function createProductFeature() {
-  const state = reactive({
-    form: {
-      name: '',
-      description: '',
-      status: ProductStatus.DRAFT,
-      price: 0,
-    } as ProductInput,
-  })
-
-  const create = async () => {
-    console.log('create product')
-    await API.graphql(
-      graphqlOperation(createProduct, { product: state.form })
-    )
-  }
-
-  const statusOptions = [
-    {
-      value: 'Draft',
-      key: ProductStatus.DRAFT,
-    },
-    {
-      value: 'Active',
-      key: ProductStatus.ACTIVE,
-    },
-  ]
-
-  return { ...toRefs(state), create, statusOptions }
-}
-
-const { form, create, statusOptions } = createProductFeature()
+const { form, create, statusOptions } = useCreateProduct()
 </script>
